21：36

+++ jywz/src/main/java/xft/workbench/backstage/type/biz/TypeBiz.java
@@ -102,10 +102,10 @@ public class TypeBiz {
 	@Transactional(propagation=Propagation.REQUIRED, rollbackFor=Exception.class)
 	public void deleteMaterial(Integer id) throws Exception {
 		
-//		boolean flag = checkTypeService.checkCodeStatus(id);
-//		
-//		if(!flag)
-//			throw new KPromptException("物资类型被使用了，不允许此操作！");
+		boolean flag = checkTypeService.checkMaterialStatus(id);
+		
+		if(!flag)
+			throw new KPromptException("物资被使用了，不允许此操作！");
 		// 删除物资
 		typeDao.deleteMaterial(id);
 	}
